promoting frequency from float to double

+++ irobot_create_toolbox/include/irobot_create_toolbox/vector_publisher.tpp
@@ -21,9 +21,9 @@ VectorPublisher<T, V>::VectorPublisher(std::string publisher_topic, std::vector<
 
   std::cout << subscription_topics[0] << std::endl;
 
-  const float frequency{62.0};  // Hz
+  const double frequency{62.0};  // Hz
   timer_ = this->create_wall_timer(
-    std::chrono::duration<float>(1 / frequency),
+    std::chrono::duration<double>(1 / frequency),
     std::bind(&VectorPublisher::publisher_callback, this));
